Possibly the final version??? read notes for bugs
right few bugs to be aware of in the ZIP version im uploading, when you load a game on level one, for some reason you can walk through a handful of walls. on level 6, the if the enemy walks into you, you dont die until you step into a wall? and on level 7 the teleporter teleports you sideways so try not to use that level